prepare_joints Function and more prep work
Added the prepare joints function, in a chain before the solver, im
hoping that the apply_deffered function gets put in the middle, i'll
find out soon.

prepare function firstly queries all joints that have parents and adds
JointParent to them, which in turn adds itself as a child (and the
JointChildren component too if it's not already there) to the parent
joint.

I like the idea of using a relationshiptarget with the joints, as
opposed to making a giant sorted vec of joints on the base, along
with a reference to the end effector (that was how my last 'test run'
implementation went). It means that everything is more integrated into
the ecs, that means i can do regular queries for joints, which can
be standalone components. Meaning i can create a rotation constraint
component later down the line which can be standalone too. Using a
One-To-Many relationship also means, as the name suggests, that bones
can have multiple children and joint chains can have multiple end
effectors, as the paper puts it. This is key to making a full human
body joint chain, starting from the hip and ending at the hands, feet
and head.

Based on what is described in the papers, i'll need another component
called a 'sub-base' which will sit on joints with more then one child.
More about the specifics in another commit.