Added JTransform, Config, SubBase, solver setup. This Commit, I added JointTransform, IkGlobalSettings, SubBase and a solver setup function that gets the branch nodes needed to start the forward reach:
JointTransform is a transform that will exist solely in joints,
containing an Affine3A, and will be used for the following:

- Handling all the unpcoming IK math.

- access to modifiable, global transforms that will be updated at the
end of the solve system.

- easy sync back to regular transform.

I had a similar system in place in my previous test run using bevy 0.15,
using a vec of fresh Transforms, which was then used to update the real
ones at the end of the solve system. i could get away with this because
i could know the order and amount of transforms ahead of time, which
definitely made it easier. Unfotunately, that isn't viable this time
because i no longer know how many joints there will be ahead of time,
and order is difficult, with the introduction of sub bases and multiple
end effectors.

IkGlobalSettings is a new Resource that currently houses the global
number of iterations to do. It used to be in the plugin but that proved
painful so i removed iterations from there and added them here. Not much
else to say really.

SubBase is a Tag Component that gets added to joints with more than 1
joint child, this situation is found in joint chains with multiple end
effectors, an example of this could be a tree with branches, or more
likely, A human skeleton rig which branches out into feet, legs and a
head. This SubBase Tag may or may not be around in the future, because
it probably isnt very useful when i can just check if a joint's children
component value length is more than 1, and also have access to said
children.

finally, a solver setup function, called forward_reach_setup was added,
this runs before any iterations, and within the base loops (each joint
chain is seperated by all having 1 base). it loops through each joint
going down, adding the most current ones to a vec of entities. when
a joint has multiple children, it adds them to the vec to be checked,
while replacing the original entity in the vec with the first child,
there is probably a better way to do this then the way i did it, with 2
while loops. but it's good enough for now and i'll update it when i find
a better way to do it. i also might move it to it's own function before
the solver body, attach the vec to the base component. and only target
added bases, then when the solver finds more end points then what it
expects, or the end points have changed, it tells the system to update
the vec at the end of the solver.

The Simple Example was also modified, to make sure multiple branches was
working, may be reverted back.

I also had an idea to make another One-to-Many relationship target
between the Base and BaseJoints. this way, i could handle endpoints
easier, as described above in the comment. In Fact, i could just replace
the BaseJoint and EndEffector Joint Components with relationship targets
both, this would allow an endpoint to target multiple joints. while the
base can target multiple endpoints as opposed to multiple joint chains.
you should just make another base entity.