Added Check Reachable Added The CheckReachable function, whose purpose it to check if the end joints can actually reach to the targets. this is an important optimization, it means we can run a special and cheap one iteration reach that will automatically give us the optimal outcome.
Right now it returns a bool, but ill change it to be a vec of entities
(sub bases) to ignore in the chain. and add an exit out if the length of
the vec is equal to the length of the sub chains, that way a chain can
be partially unreachable.

The forward reach setup was also modified by calculating the total chain
length for each subchain.