Changed check_reachable, forward_reach_setup Changed check_reachable function to output a vec of a tuple of 3 entities; the endpoint, the subchain, and the end effector entities, in that order.
Also modified the forward_reach_setup function, before, everytime we
found a sub base, the sub base entity would be added only once overall
to the sub chain vec. I thought this might bring up issues in the
future, and i also thought it was confusing. Now when a subbase is
found, the subbase entity is added to the vec multiple times, depending
on the amount of children added to the endpoint vec. This way each
endpoint has a clear root. However, it might be a good idea to instead
add the child as opposed to the parent. I don't know what is the better
solution at the moment, so ill leave it for another time.