Finished reach_out_full and added sync function Firstly, i finished the reach_out_full function, everything has been tested and works correctly. After the first "successful" test, there was an issue where at the first joint of a branch, the joint would be in the incorrect place; half way up the first child joint. This is because in the simple example, with the cones for bones: The middle of the cone is positioned at the local transform center, so it would go too far up the direction of the parent joint. This was solved by making the next child base be half way up the current joint. Then, when the child joint comes around, set its position to the base, then align it, then get the new local up, after the align, and multiply that by half the current joint length, and add it to the current position. This ended up working, but it wont for joints that arent positioned at half way, perhaps ill have to add a bool in the joint struct.
Going along with this, i also added code at the bottom of the solve
function which recursively syncs the joint transform and the real
transform. This is done by using TransformHelper to get the currentmost
global transform of the target parent, then get the inverse affine of
that, and multiply it with the joint transform affine.

In my previous "test run" of a bevy based fabrik solver, syncing
the transforms was by far the biggest difficulty, firstly because
syncing back and forth between global and local is a pain in the bum,
and secondly i didn't really know what i was doing. These Issues are
alleviated by:

- understanding the correct way to sync back a transform.
- me knowing what im doing.

i also learned that i can very easily sync the affine joint transform to
the real local one by multiplying the affine of the joint transform with
the inverse affine of the parent global transform, no aligns or whatever
else needed. of course the solution looks slightly different when there
isn't a parent, but it's not much more difficult.

i also changed the sub base target in the setup function to be the child
instead of the parent, it just makes more sense, and i use the child in
the reach_out_full function.

i noticed that the length calculation is going wrong some way somehow,
being a bit longer than it should be, not sure what it is but ill add a
commit for it when i find it.