Partially Fixed Forward and backward reach partially fixed forward and backward reach. changed both functions from the ground up, instead of using a weird vec system we use hashsets and a vec / vec dequeue. There was also the silliest mistake ever where i was giving the forward reach setup the unreachable end points only, which was causing it to not work.
Apparently i got something wrong somewhere, because when only one of
the targets are in vision. It points towards the unreachable for some
reason. However, if both targets are within reach, then it does a
strange spinny dance, which while fun to watch, is not really what i was
going for.