Removed SubBase, Fixed Algorithm issues + more This Commit I Fixed The Algorithm issues i was talking about, where objects would grow and form in strange ways, this was due to scale being incorporated, changing it to not take any scale in, and not return any either, all strange issues were fixed. Although i feel a bit bad about getting and setting it as 1, i think getting the current, and then returning what it is a better idea, or just not setting the scale whatsoever. This was actually fixed in the last commit, but i forgot to talk about it.
Secondly, i removed the sub base component, i figured out it's just
completely useless so i removed it. The behaviour for 1 child or more
than 1 is the same right now anyway.

I also added back the End Effector joint setting functionality, that is;
The Bool to make the joint be in the center, and the Bool to make the
joint be set to the rotation of the End Effector.

Now that I've done some polishing, i think ill try adding rotation
constraints again, in the form of Swing And Twist i failed this last
time, and went for some euler cop out method, so i want to try getting
it right this time.