Started Constraint Overhaul: constrain_forward Started The Constraint Overhaul after a day of research and playing around with Twist Swing decomposition.
Firstly:

Rotation Constraint Component Was Redone, now has 5 vecs and a
Quaternion; Decomposition direction, The Swing min and max, as well as
the Twist min And max, and an offset quaternion. The offset quaternion
probably won't last much longer since you can set the vecs to be
whatever you want, the Decomposition direction is for the direction that
the twist is in. so you could make it be in the z direction and almost
completely constrain sway, which would give you a sort of hinge joint,
although you could just instead constrain one of the axes. The Mins,
and Maxes are what they say they are, they clamp the twist and swing
respectively, converted to a scaled axis.

Old Constraint code was commented out (for now), and in it's place
is the first of two functions: constrain_forward, which takes the
child's rotation constraint component, and uses that to constrain the
direction based off of the values, then resets the value of p_i based
on the new direction. Due to the way it works, the first and last
joint don't get constrained. But it seems to work in all situations
other than extreme ones, and has more control compared to the last.
However, right now it only takes into account the first child with
a RotationConstraintComponent. In reality, there should be a sort of
average direction, and the function be run for all with the component.
So that there won't be any bias towards the constrained joint.

Next up, i need to add the second function: constrain_backward, which will do a similar thing to what the current constrain does.

After fiddling around with twist swing in a seperate project. i figured
out that alot about it, like that the order of so many things, just
by changing out = swing * twist; into twist * swing, will instantly
invalidate the transform, make it spin uncontrolably for some reason.
Same thing happens with the parent conjugate * theoretical. It's just
ridiculous. I think having swing and twist round the wrong way is why my
last two attempts just flat out failed.