Added back constraint file + solver work Figuring out the solver stuff has been quite difficult, however my old code has been a big help for what is probably the first time, which i'm glad about. I also realised i was fretting over nothing about parallel access of the queries. Since i realised that as long as i dont do anything mutable with them, i can use them for the inter parallelism or whatever it's called. Because they already have to be able to support being accessed in multiple different threads thanks to bevy forcing systems to be parallel when it's possible.
I also spotted an issue with the pre rewrite code with doing offsets,
which are very important for a lot of different types of rigs.

The issue was that i was pulling the parent rotation instead of
the child rotation during the forward pass, causing the joint to be
missplaced, which is corrected by the backward pass, but then the cycle
repeats, causing those jiggling issues that were seen with any offset
with a magnitude greater then like 0.05, the backward pass was probably
correct but because the front one wasn't, it was causing the jiggling.

I'm sure there is a similar issue plauging the joint constraints, which
were honestly no good to begin with, so i'll have to figure out how to
fix that too.

Funny that this is my 3rd serious attempt.. the first one used bevy
0.15, which was before 0.16 and the relationship reworks, which was
very limited (no branching joints or constraints, the latter of which i
tried and failed at). then came around 0.16 and the second rewrite came
in (the previous one) and it was a huge leap forward however there were
some issues and i ran out of momentum.

And now, 0.17 is almost out, which ill be updating this project to, i
don't even know of any of the new features other then dlss support and
the guaranteed meshlet improvements which is something i always like to
see. and as for the project, well i know how to do rotation averaging
properly now. with some covariance eigenvector method, and i get some
more motivation and the benefit of hindsight.

Next on the agenda is reimplementing utils.rs and all the functions,
because i need (really want) some of the functions there to continue
with the solver work. I may also add a todo.md file, seems like a good
idea, some way to know what to do, first in the short term, then longer
after that. Till next time.