Finished Forward Reach finished the forward reach (i think), there will probably be bugs, or maybe even a deadlock but nothing in particular looks particularly no good to me, this function took ages because i wasted a lot of time trying to figure out the way of forward joint constraints in another git project. ditched it and went back to this to get it done.
also read up on deadlock on the expanded fabrik paper, and it looks
like the way to do that is to terminate early if the end joint doesnt
reach, which could apply for individual ends, by passing in a cloned
local version of the bookkeeper, then after the forward and backward
passes are done, do a cull pass, where you check each end and see if
its reaching / able to reach, and if it cant (check maybe by accounting
for joint limits and going back a joint or two and checking them too),
remove it from the local ends vec, so it doesnt get checked in the
future.

before i do that, i need to do the back pass, which should be easier
to implement.