Migrated to bevy 0.17 + rotation average rework Migrated to bevy 0.17, pretty much nothing changed except for a naming convention. but i intend to release it as a crate when it's ready.
I reworked the rotation average function, rust write of a rotation
average function found on the blog post of one Daniel Holden, according
to the post. the function is order independent and doesn't have any
harsh jumps when a quaternion changes hemispheres. i also get to use
weights, which leads to a modification of the constraint struct, adding
a value at which the other joints should be prioritised in the average.
hopefully people will be able to use it to make certain joints, like
maybe spine joints for example, less effected by the rotation of many
children.

Also added the main joint to the rotation list. and for now im also
constraining it if there is a parent. it's possible that the function
im using may not be the right one (constrain_forward) but we will see.
i also may change the main rotation to not be it's constrained rotation.
but instead it's identity. if i do then i might add another variable
to change the weight of the main joint too. once i can properly test
things, i will.

Worked more on backward pass, it seems to be coming along nicely.

A want to get this out before the 13th, and if not then the 20th of
october. so i'll have to pick up the pace.