Worked on Backward pass The Backward pass is all but finished, now whats left is to test and see how things are. last commit, the backward pass had a deadlock, and i decided to just keep writing and when i tested it now, the commit disappeared. I suppose i should be thankful, but anyway.
The night of the previous commit, there was some power outage or
something, and it killed my main pc's power supply which is really
annoying, a new one has been ordered but it wont get here for a while.
I'm really hoping that it's just the power supply, because if it's the
motherboard aswell (which it could possibly be), things are going to
be shit for a couple months since i dont have new motherboard sort of
money.

Hopefully things will be up and running in a commit or two.