Debugging more debugging, things are working but not working, also the new PSU came in today, and... the pc is still dead. FML.
Meaning i'll have to get used to this laptop which sucks but oh well.

i think ill revert back to the pre september rewrite commit and bring
the key changes from current to there, stripping everything out was
helpful to relearn what i was even on about, but now I'm sort of
regretting not having a working base upon which i can improve.