Fixed more issues fixed the issue where the joints weren't moving.
The issue was that change detection bypass of the transforms in the sync
stage was stopping the global transform calculation. This made it so the
joint wouldn't viusally sync up until given a nudge from the inspector.

Currently there are some joint possitioning issues, almost certainly
due to some funny handling of rotations. specifically the own
transformation.

I was also thinking of refactoring the weight variable. instead of
affecting child joints, it effects itself when worked upon by parent
joints.