Partially fixed Rotation Constraints issue + Other Firstly the issue with constraints has been fixed, where joint positions were strange, however the issue still partly exists.
Im fairly certain that it comes from the way i decided to try doing
things in the backward constraint function, at the start, it takes
all the base joints, and puts sets them to the bases, but leaves the
rotation the same, when it should be following the algorithm the same as
the rest.

Even though in the algorithm, the first point is always set to the base
at the start of the back pass. In our implementation, we have bones,
which share the top point with the next joint's bottom point. I Did not
account for this, which is why there is an issue.

I say that, but i haven't actually gone and fixed the issue yet. That's
what next commit will be about.

I also reworked weights. Now, instead of dictating the weight of all
other joints on the current one, it dictates the weight of the current
joint on all others. This will make it easier to prioritise specific
joints in rotation. With the old system, say on the top spine joint of
a humanoid rig, setting the weight low would have just made it stick
closer to it's current rotation, rather then towards one specific
joint, which can be done now, say the neck joint instead of the shoulder
joints.

Once again, i haven't actually tested it yet so my theory could be
wrong, and i, talking out my ass.