Fixed Bottom joint issue Fixed the bottom joint positioning issue, where the bottom joint wouldn't rotate correctly. Achieved by correctly running the algorithm on the first joints.
While i was testing the changes, i noticed a new issue which is fairly
serious, and will be the target of the next commit. When moving the end
along a 3rd vector (say along the z axis), after changing the position
of the other two translation axes (and hence, the translations of all
the joints): The App crashes reporting a panic of a non normalized
vector being pushed through dir::new_unchecked.

It is originating from the quat / affine function align() which was
copied from Bevy's Transform component. i use it to align the rotation
of the joint at the end. While this definetly didn't happen in the
Pre Sept rewrite, i can also pin part of the cause on the new way of
handling the collection and re syncing of transforms.

An issue that plauged the Pre Sept Version was errors on the affine
usually thanks to scale accumulating in a catastrophic way, showing
itself as the contortion of the transforms, initially, whereafter
causing the mesh to explode and crash the App. The way i fixed it before
was to exclude scale from the affine entirely, both at collection and
sync. But this obviously means that scale gets invalidated.

The way i plan to solve this is with a rework to JointTransform,
removing the affine, and replacing it with 3 seperate Scale, Rotation
and Transform variables, the same as a normal Bevy Transform Component.
I Thought that using affine would improve things both ergonomically
and perfomance wise for the algorithm. In the ergonomical sense,
while things were improved in some ways, somewhat, the rotation had
to be calculated at various inconvenient points throughout the solver
functions. As for the performance side: I don't actually know, this is
because i haven't done any real benchmarking at all. Making desicions
on implementation, as well as making major changes based on the guise or
idea of improved performance, is a bad habit i continue to make time and
time again. A recent example is with the September rewrite, changing the
functions to be parallel. Atleast with that, its pretty much guaranteed
that performance will improve with a large amount of chains. obviously
with degredated performance compared to non multi threaded with low
counts.

On the topic of performance and such, i also plan on adding a
multithreaded feature to the cargo. Also with the soon-to-be bisected
Joint Transforms, the idea of Simd becomes more feasible, if i ever get
around to that, in which case ill need to figure out what platforms i
want to support, and more importantly, how to even do such a thing.