Migrated off of Affine Joint Transforms Migrated off of affine joint transforms into a bisected Translation Rotation Scale Style transform, the same as bevy, so that scale can be retained, and such.
While the rework was successful, and fixed the App crash issue when
moving on more then two axes, a new issue has shown itself, or atleast
become more apparent. When moving the end joint, the 3rd joint in
the chain (second from the top), jitters, seperating from the top
joint's base, before reattatching. The effect occurs on both 2 and 20
iterations.

Not sure what the issue is at this moment, but the next commit will
hopefully fix it.

Also tested out offsets, and once again, they're broken. When reviewing
it, i was in two minds about how it should be done, and im pretty
confident that it should be about the parent's rotation. However, i need
to figure out where to apply that and how.