Fixed Instabilities Fixed the instability issue that were occuring in certain positions, did this by:
- Changing the rotation that would be constrained and sent to the vec
to be the potential rotation of the main joint (the direction by doing
top - bottom normalized, convert rotation with align), instead of the
rotation of the child joint.

- If the the len of children, + end effector is less than or equal to 1,
just use the first entry of the vec, no trying to do rotation averaging.

- Also removed the first self entry that it was doing, before i did
this, the end effector wasn't working correctly. now that it is, it's
fine.

- Other miscellaneous fixes.

Pondering whether i should add the first self entry back or not, but i
don't think so, and if i do, it would be the identity rotation, not the
current one.

Next on the list; Optional final global transform syncing, enabled with
the global config incase you have a recursive hierarchy, like the one
that was causing issues a commit or two ago. Should fix the problem
entirely.

Proper offsets that actually work. i removed a bunch of hodgepodge
crap from the forward pass that didn't make sense. I Think ill collect
the offsets seperately, then figure it out at the end. The only reason
things didn't go to shit with small offsets before, is that the back
pass offset handling is mostly correct.

Some better Examples, testing, and a humanoid example. Should be ready
for release after that.

Also, this will hopefully be the last commit from this laptop, atleast
where i'm forced to use it, other then the last part which hasn't come
in yet; the CPU, the computer is pretty much finished. may have to wipe
the drive and start from scratch but we'll see.