Early exit via satisfactory tollerance value added Readded the early exit via satisfactory tollerance, during the backward pass, it adds the up the difference between all the effectors and their respective joints, then at the end, once it's finished, it compares it with the previous, then sets the old to the new, and if the difference between the two is less then the minimum tollerance, it breaks out of the loop.
I also tested the standard configuration with rotation constraints,
and while it is perfectly stable, it isn't being correctly constrained,
which means that i need to make a backward constraint function, which
should be easier.

Also need to do the toggleable global transform setting function at
the end, as well as sorting out offsets, the former should be pretty
easy, offsets on the other hand are very tricky. even more so then
constraining and averaging of rotations.