Added proper Transform Hierarchy support finally got around to adding proper transform hierarchy support, that is when you have a hierarchy of joints, (where each joint is child of the one before) the correct transform is output. when this case occured previously, updating the local transform of a joint would not count for the local transform of it's parent, which may or may not have be outdated compared to what it should be, this random outdatedness would create errors in the position and rotation of the transform, compounding the error in some cases, with a larger error with more joints. It would always sort itself out a couple frames later naturally, but it would still create very visible errors when moving the ends or bases, especially a large amount.
The way this is fixed is by setting the global transform to what it
should be based on the joint transform after the transform propogation
step. this completely elimintates the problem, and the system can be
toggled with a bool in the global settings resource. it felt a little
daunting to get this done but was actually a 5 min fix. I'm also
pondering if i should get rid of the other system entirely, not that the
main one doesn't have it's merits of course. although it is incorrect in
certain situations.

Next up, and the last major thing to fix before making examples and
releasing, is that the rotation constraints are broken, i'm still not
sure what the issue specifically is, but i atleast think it isn't todo
with using scaled axes specifically, although it might be a better
choice to use euler angles instead, which is a little funny, but i don't
know yet of course, the current idea is that it has something to do with
the way things are being clamped. but we'll see.

My PC is fixed, or really i should say i have a new one now, because the
only old thing about it is the graphics card and the storage, it was a
huge pain to get running, and still has issues, but overall is running
better and faster then both the old pc and the macbook with linux on it,
which coincidentally has stopped working, (atleast linux). I'll have to
sort that out too, but all progress will be slow while i have Univerity
to attend.