Fixed Rotation Constraints and Offsets The issues with the rotation constraints and offsets have been fixed in what was pretty much all at one fell swoop, which im pretty proud of myself for.
The Backward pass was changed, so that instead of doing the the correct
algorithm in the backward pass, we very simply just set the current
bottom point to the previous top point, without trying to make anything
point anywhere else, which preserves the rotation constraints.

Offsets now work(*) too, and consequently the joint halfway boolean and
joint centre boolean no longer work correctly. this is a tradeoff i am
willing accept. the whole idea of halfway is kind of silly anyways, i
forgot to get rid of those related variables but i will next commit.

* Offsets are almost what i imagined them, with the one issue where a
joint wont properly connect to it's parent if there is an x or z offset,
this is either some bug / oversight in the code on my part, or having to
do with not actually making the translation the bottom point, but trying
to convert between the raw translation and the bottom point, it will be
the main focus next commit either way.

Thanks to university, i have had less time to sink in to this project
then i would have liked, and the time i'll be able to put into this in
the future is shorter then hoped as well, but at the same time the break
i got from this project was needed to solve these two large issues.