Added anchor offsets Anchor offsets are the new offset in the joint component, they are an offset of the base (anchor) of the joint and are relative to the parent joint, this is opposed to the old offset, now named visual offset (since it has no effect on the kinematics of a joint), which was relative to itself. This means that the new anchor offset will do nothing on a base joint, because there is no parent joint to base it off of, i could have done something with the base (and i might in the future), but i chose not to in the end.
This problem was unusually annoying to figure out, but it's done, and
unlike the rotational constraints, there is no more work to do on it
to improve it, I'm also keeping the old offsets around, now called
visual offset, because it's required for the simple example with the
primitives, and because of the way FABRIK does things, you can't really
transform a joint on it's own, which woud probably be fine to work
around with the primitive example by putting the meshes in a child
of the joint, im not sure this would work with any skinned mesh /
hierarchical bone system.

I also deleted algorithm md because it's not useful to me anymore and
it also doesn't really apply to this project anymore the way it has
deviated so much.

Next up is eliptical joint constraints, also need to figure out how to
get back out twist constraint, might be the constraint of yet another
unit vector, or some form of rotation constraining (probably not).

I also want to restore the backward pass to actually do backward pass
things, rather then just setting positions, i realised that i need to
constrain rotations here aswell, which shouldn't be a problem anymore
now that things are working correctly.

After that, is probably going to be the making of some extra gizmo
plugin, then ill make a proper skinned mesh demo, then i'll hopefully
release it after that.