Started Gizmos, fixed backward pass backward pass is ready now, also added a gizmo plugin which is good.
Started off with the humanoid example but it's going shit. I have a
feeling it has something to do with the identity rotation not meshing
right with the constraint stuff, it just looks wrong, plus there was
some sort of global transform issue that i was seeing a while ago. ill
have to dig into the constraint stuff but tomorrow is christmas so it'll
have to wait.

I'll probably have to treat the identity like the same as i do the
visual offset.