Removed pinhead and humanoid_primitives example First of all i removed pinhead, he was pretty ugly. Good riddance!
Secondly, the humanoid primitives example was removed, because it
exposed some pretty serious issues with the solver, specifically the
Rotational Constraints and also the rotation averaging.

Now before you start (if it's someone other then me reading this):
Rotational Constraints again?? I thought it was fixed the last time. And
for maybe the 5th or 6th time now; no, it's not as fixed as i thought
it was, again, for like the 5th time. To be fair, it's not that bad at
all, and is by far the best Rotational constraint i have done so far.
It works great in single End Effector chains, which is perhaps the bare
minimum, but it breaks down when we get into chains with multiple end
effectors.

This is almost certainly caused by the Quaternion Averaging function
having issues, Along side other rotation stuff. That Quat Average
Function is probably broken because:

a): It's not correct (possible)
b): I'm not unrolling Quaternions correctly (most likey and a bigger issue)

Quaternions have an unsightly side effect of having more then one
way to represent the same rotation, and also a quaternion has change
hemispheres (on a graph) when it crosses past +180 degrees, it jumps to
the negative hemisphere of -180 degrees. This is fine by itself, and in
specific (non animation related) examples. Because the 2 rotations on 2
different sides of the hemisphere can represent the same rotation.

However if we are trying to, for example, average (^^ up there!!) a
number of rotations: A rotation that was in one hemisphere, but is now
in the other, now has to travel the long way around, which will produce
a different output in our rotation function, resulting in a different
rotation.

This was probably also affecting a number of different parts of the Ik
system previously, such as the old rotational constraints, or the swing
twist decomposition, or probably other stuff that's in there now, like
maybe converting the two directions into a quaternion.

When i was having that issue of the chain being fine when it was only
moving on two axes, but then breaking and freaking out when it moved in
a tiny bit on the 3rd axis, was almost certainly caused by quaternion
discontinuities and hemisphere jumping, because the raw readouts of
the quaternion were jumping all over the place, probably because one
joint would flip hemispheres, then that would mess up the rotation
constraint big time, which would mess up the next joint, causing it to
jump hemispheres even, etc. causing a death loop of pain and hemisphere
jumping.

The Obvious message im trying to convey here is that i need to start
actually using Quaternion unrolling. this will hopefully bring back the
old rotational constraints, and improve the whole situation.

The alternative is to use a 3x3 Matrix for Rotations, which would be a
bit of a cop out to be honest, and obviously has it's own downsides.

According to The orange duck (the blog where i read basically everything
i needed to know about joints, rotations and IK except for the FABRIK
algorithm), the Mat3 is good because it's continuous full stop, however,
it apparently can't represent rotations beyond +180 or -180, as opposed
to an unrolled quat, which can do the full -360 <-> +360:

"Like rotation matrices, unrolled quaternions will not grow larger and
larger in value as you spin them around multiple times - but unlike
rotation matrices, quaternions can distinguish between rotations of
more than +180 or less than -180 degrees (up to +360 or -360 degrees, at
which point they loop around again)."

^^ unrolling rotations
https://theorangeduck.com/page/unrolling-rotations

other then all that, there was a bunch of QOL stuff:

- Defaults for End Effectors and Bases

- Cleanup of the simple example

- Introduction of the multiple ends example which suffers from all
the issues described above, but significantly less so (to the point
of stopping bevy from panicking from NAN rotations) when there are
rotational constraints which is perhaps a bit ironic.

- Gizmos settings resource

- Usage Example in the README

- Removed the inspector egui and bevy egui crates, to prepare for crate
release (not happening for a while with these issues)

- Refactored the Rotation Averaging function
to be a bit closer to the original inspiration
(https://theorangeduck.com/page/quaternion-weighted-average)

Next up is to put out a demo of the single end effector chain to the
discord.

After that, i don't really know, figure out the unrolling, im not
acutally sure when and where i should be unrolling or how i should start
out to begin with.

There is probably more small stuff but i can't think of anything right
now.