Quaternion Unrolling Added Quaternion Unrolling, both prior to the solve passes, and within the forward and backward passes.
Quaternions have a frustrating side effect known as 'double cover'.
This means that one visual rotation can be represented by two different
quaternions, one being the short way around, and the other the long way
around, say +90 degrees and -270 degrees. Each value sits on opposite
sides of the hemisphere of the so called quaternion rotation space.
If you are for example, constaining a quaternion, the long way around
might be far oustide of the constraint bounds, forcing it's way back
in, almost certainly resulting in a different rotation compared to a
short way around equivalent. This example is applicable to pretty much
everything in this project that has something to do with rotations,
including the quaternion average function.

This commit, i firstly added a quat unroll function which runs before
all the backward and forward passes in the solve function, it traverses
the joints, starting at the base joints, then going up, identically to
the backward pass. it runs the following code on the joints:

```rust
if let Some(parent) = parent_maybe {
    let parent_transform = bk.joints.lock().unwrap().get(&parent.0).unwrap().1;

    if main_transform.rotation.dot(parent_transform.rotation) < 0.0{
        main_transform.rotation = -main_transform.rotation;
    }
} else {
    main_transform.rotation = quat_abs(main_transform.rotation);
}
```

Basically, if there is no parent (a base joint) then it runs the
quat_abs function, which checks if the w value on the quat is negative,
if it is, then it inverts the whole thing. this forces the initial
rotations to all be the short way around. If there is a parent, then it
gets the dot product of the two quaternions (as if they were both Vec4),
and if the dot is less then 0 (more then 180 degrees away), it flips the
hemisphere of the rotation.

This functionality is also copied throughout the rotation averaging
function and the forward and backward passes.
pretty much all rotations made on the spot are made the short way
around(quat_abs), and just before setting the final rotation in both
passes, we take the dot of the old and new, then flip the new if it's on
the wrong hemisphere.

All of this works out to almost completely fixing the multiple end
effectors example, left with a very slight twitch in the joint below the
sub base when the two end effectors are going in different directions,
and being beyond the joint's reachability. There was no change in the
simple example.

There was one other change that improved the stability of the multiple
end effectors example: The EE handling was moved above the parent
child handling, and the child anchor positions were caluclated with
a loop before the main children loop to get the average top point
before rotation constraining, then the direction was changed to be
the difference of the new average point and the base, which was used
for all children, instead of using the child anchor as the top for
each direction to be constrained. This made a noticable difference in
stability when there were multiple joint children.

Outside of that were some small changes and fixes not worth noting, i
did find a bug with the end effector defaults, if i actually added them,
with the joint entity set to placeholder, even though it's meant to be
handled by the joint hooks, it wasn't and instead it caused a crash.
not sure how to fix this outside of making the joint on ee and base be
optional with Some().

Now that multiple end effector example is basically perfect, and
quaternions are way less scary and complicated to me now, i'll revisit
the old rotation constraining with swing and twist. it's much more
versatile then what i have currently. After that, ill try give the
humanoid example another shot.